+++ src/ragger/__init__.py
@@ -17,7 +17,3 @@ try:
     from ragger.__version__ import __version__  # noqa
 except ImportError:
     __version__ = "unknown version"  # noqa
-
-from ragger.logger import init_loggers
-
-init_loggers()