+++ src/ragger/conftest/base_conftest.py
@@ -7,7 +7,7 @@ from ragger.backend import BackendInterface, SpeculosBackend, LedgerCommBackend,
 from ragger.navigator import Navigator, NanoNavigator, TouchNavigator, NavigateWithScenario
 from ragger.utils import find_project_root_dir, find_library_application, find_application
 from ragger.utils.misc import get_current_app_name_and_version, exit_current_app, open_app_from_dashboard
-from ragger.logger import get_default_logger
+from ragger.logger import get_default_logger, init_loggers, SHORT_FORMAT, DEFAULT_FORMAT
 from dataclasses import fields
 
 from . import configuration as conf
@@ -49,6 +49,12 @@ def pytest_addoption(parser):
                      default="default",
                      help="Specify the setup fixture (e.g., 'prod_build')",
                      choices=allowed_setups)
+    parser.addoption('--silent', '-S', action='store_true', default=False, help='Limit log traces')
+
+
+@pytest.fixture(scope="session")
+def silent(pytestconfig):
+    return pytestconfig.getoption("silent")
 
 
 @pytest.fixture(scope="session")
@@ -151,12 +157,19 @@ def pytest_generate_tests(metafunc):
         metafunc.parametrize("firmware", fw_list, ids=ids, scope="session")
 
 
-def prepare_speculos_args(root_pytest_dir: Path, firmware: Firmware, display: bool,
-                          cli_user_seed: str, additional_args: List[str]):
+def prepare_speculos_args(root_pytest_dir: Path,
+                          firmware: Firmware,
+                          display: bool,
+                          cli_user_seed: str,
+                          additional_args: List[str],
+                          silent_speculos: bool = False):
     speculos_args = additional_args.copy()
 
     if display:
         speculos_args += ["--display", "qt"]
+
+    if silent_speculos:
+        speculos_args += ["--silent"]
 
     device = firmware.name
     if device == "nanosp":
@@ -214,9 +227,14 @@ def prepare_speculos_args(root_pytest_dir: Path, firmware: Firmware, display: bo
 # Depending on the "--backend" option value, a different backend is
 # instantiated, and the tests will either run on Speculos or on a physical
 # device depending on the backend
-def create_backend(root_pytest_dir: Path, backend_name: str, firmware: Firmware, display: bool,
-                   log_apdu_file: Optional[Path], cli_user_seed: str,
-                   additional_speculos_arguments: List[str]) -> BackendInterface:
+def create_backend(root_pytest_dir: Path,
+                   backend_name: str,
+                   firmware: Firmware,
+                   display: bool,
+                   log_apdu_file: Optional[Path],
+                   cli_user_seed: str,
+                   additional_speculos_arguments: List[str],
+                   silent_speculos: bool = False) -> BackendInterface:
     if backend_name.lower() == "ledgercomm":
         return LedgerCommBackend(firmware=firmware,
                                  interface="hid",
@@ -227,7 +245,8 @@ def create_backend(root_pytest_dir: Path, backend_name: str, firmware: Firmware,
     elif backend_name.lower() == "speculos":
         main_app_path, speculos_args = prepare_speculos_args(root_pytest_dir, firmware, display,
                                                              cli_user_seed,
-                                                             additional_speculos_arguments)
+                                                             additional_speculos_arguments,
+                                                             silent_speculos)
         return SpeculosBackend(main_app_path,
                                firmware=firmware,
                                log_apdu_file=log_apdu_file,
@@ -242,9 +261,10 @@ def create_backend(root_pytest_dir: Path, backend_name: str, firmware: Firmware,
 @pytest.fixture(scope=conf.OPTIONAL.BACKEND_SCOPE)
 def backend(skip_tests_for_unsupported_devices, root_pytest_dir: Path, backend_name: str,
             firmware: Firmware, display: bool, log_apdu_file: Optional[Path], cli_user_seed: str,
-            additional_speculos_arguments: List[str]) -> Generator[BackendInterface, None, None]:
+            additional_speculos_arguments: List[str],
+            silent: bool) -> Generator[BackendInterface, None, None]:
     with create_backend(root_pytest_dir, backend_name, firmware, display, log_apdu_file,
-                        cli_user_seed, additional_speculos_arguments) as b:
+                        cli_user_seed, additional_speculos_arguments, silent) as b:
         if backend_name.lower() != "speculos" and conf.OPTIONAL.APP_NAME:
             # Make sure the app is restarted as this is what is requested by the fixture scope
             app_name, version = get_current_app_name_and_version(b)
@@ -314,8 +334,12 @@ def pytest_configure(config):
         "needs_setup(setup_name): skip test if not on the specified setup",
     )
 
+    if not config.getoption("--silent"):
+        init_loggers(DEFAULT_FORMAT)
+
 
 def log_full_conf():
+    init_loggers(SHORT_FORMAT)
     logger = get_default_logger()
     logger.debug("Running Ragger with the following conf:")
     for field in fields(conf.OPTIONAL):