+++ CHANGELOG.md
@@ -5,11 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.28.0] - 2025-03-04
+
+### Added
+
+- Add -S/--silent option to limit logs (and forward the option to speculos)
+
 ## [1.27.1] - 2025-02-26
 
 ### Fixed
 
-- Fixed backend being instantiated even when the test needs a different setup 
+- Fixed backend being instantiated even when the test needs a different setup
 
 ## [1.27.0] - 2025-02-21